bot/requirements.txt
discord.py==1.6.0
asyncpg
tortoise-orm
python-dotenv
aiohttp_cors